test: la campagne sur les quatre moteurs, et les quatre désaccords qu'elle a trouvés
Elle a payé au premier passage. Chacun des quatre aurait donné une application qui
démarre sur la machine du développeur et échoue ailleurs.

**Les identifiants étaient `int` face à un champ `Long`.** Le schéma NestJS utilisait
`SERIAL` parce qu'un nombre JavaScript ne distingue pas les deux. Passés en `bigint` :
quatre octets de plus par ligne, et pas de plafond à deux milliards de constats.

**`boolean` devient `tinyint` sur MySQL, et Hibernate le refuse — il attend `bit`.**
MariaDB, de la même lignée, passe sans rien dire. C'est exactement l'argument pour
lancer les quatre plutôt qu'un représentant : les deux moteurs que tout le monde traite
comme interchangeables sont les deux qui diffèrent.

**SQLite refuse `AUTOINCREMENT` sur autre chose qu'`INTEGER PRIMARY KEY`** — et cette
colonne est un alias de rowid, donc déjà sur 64 bits. Le type qui paraît le plus étroit
est le plus large, et écrire `bigint` par souci de cohérence produit une table qu'on ne
peut pas créer.

**Et celui qu'aucune vérification de type n'aurait vu.** Sur SQLite, Liquibase traduit
`datetime` en TEXT ; le pilote y écrit un instant sous forme de nombre et n'arrive plus
à le relire — « Error parsing time stamp », à la lecture, longtemps après une écriture
qui avait l'air correcte. Le schéma était cohérent avec lui-même. Il a fallu écrire une
ligne et la relire. C'est précisément pourquoi SQLite ne subit pas la validation stricte
mais un aller-retour : sur un moteur sans types, comparer des noms de type mesure les
choix de nommage de Liquibase, pas si la correspondance fonctionne.

Une tâche Gradle par moteur, chacune avec son répertoire de résultats : un répertoire
partagé fait écraser les précédents par le dernier, et « quel moteur a échoué » cesse
d'avoir une réponse.

Testcontainers 2 renomme tous ses modules avec un préfixe et déplace les conteneurs
concrets dans un paquet par module — les classes de même nom sous
`org.testcontainers.containers` sont les anciennes, toujours présentes et toujours
compilables, ce qui est la façon dont un build continue de les utiliser en silence. Sa
version, comme celle de Liquibase, est rendue au BOM de Spring Boot.

+++ zanshin/zanshin-core/src/main/java/com/asmolabs/zanshin/core/persistence/ApiKeyEntity.java
@@ -50,7 +50,7 @@ public class ApiKeyEntity {
     private String targetKind;
 
     @Column(name = "target_id")
-    private Integer targetId;
+    private Long targetId;
 
     @Column(name = "expires_at")
     private Instant expiresAt;
@@ -119,11 +119,11 @@ public class ApiKeyEntity {
         this.targetKind = targetKind;
     }
 
-    public Integer getTargetId() {
+    public Long getTargetId() {
         return targetId;
     }
 
-    public void setTargetId(Integer targetId) {
+    public void setTargetId(Long targetId) {
         this.targetId = targetId;
     }
